requirements.txt
# Core Telegram Bot Framework
python-telegram-bot==20.7

# Google API Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# NLP and AI
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2

# Vector Database
pinecone-client==2.2.4

# Web Framework (for webhooks and API endpoints)
flask==3.0.0
flask-cors==4.0.0

# Task Queue and Scheduling
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
jinja2==3.1.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

